fix: decouple repository mixin to break circular imports
