+++ pkgs/standards/peagen/peagen/orm/keys.py
@@ -12,7 +12,7 @@ from autoapi.v2.types import (
 from autoapi.v2.types import relationship
 from autoapi.v2.tables import Base
 from autoapi.v2.mixins import GUIDPk, Timestamped, UserMixin
-from peagen.orm import RepositoryRefMixin
+from peagen.orm.mixins import RepositoryRefMixin
 
 
 class PublicKey(Base, GUIDPk, UserMixin, Timestamped):
